refactor(api): change new json package name to avoic conflict
