+++ api/internal/handlers/login.go
@@ -7,12 +7,13 @@ import (
 
 	"github.com/benKapl/cvmaker_api/internal/auth"
 	"github.com/benKapl/cvmaker_api/internal/database"
+	"github.com/benKapl/cvmaker_api/internal/respond"
 )
 
-func (cfg *apiConfig) handlerLogin(w http.ResponseWriter, r *http.Request) {
+func (a *API) handlerLogin(w http.ResponseWriter, r *http.Request) {
 	type parameters struct {
-		Email    string `json:"email"`
-		Password string `json:"password"`
+		Email    string `response:"email"`
+		Password string `response:"password"`
 	}
 
 	type response struct {
@@ -25,45 +26,45 @@ func (cfg *apiConfig) handlerLogin(w http.ResponseWriter, r *http.Request) {
 	params := parameters{}
 	err := decoder.Decode(&params)
 	if err != nil {
-		respondWithError(w, http.StatusInternalServerError, "Couldn't decode parameters", err)
+		respond.WithError(w, http.StatusInternalServerError, "Couldn't decode parameters", err)
 		return
 	}
 
-	user, err := cfg.db.GetUser(r.Context(), params.Email)
+	user, err := a.DB.GetUser(r.Context(), params.Email)
 	if err != nil {
-		respondWithError(w, http.StatusUnauthorized, "Incorrect Email or Password", err)
+		respond.WithError(w, http.StatusUnauthorized, "Incorrect Email or Password", err)
 		return
 	}
 
 	err = auth.CheckPasswordHash(params.Password, user.Password)
 	if err != nil {
-		respondWithError(w, http.StatusUnauthorized, "Incorrect Email or Password", err)
+		respond.WithError(w, http.StatusUnauthorized, "Incorrect Email or Password", err)
 		return
 	}
 
-	accessToken, err := auth.MakeJWT(user.ID, cfg.JWTSecret, time.Hour)
+	accessToken, err := auth.MakeJWT(user.ID, a.JWTSecret, time.Hour)
 	if err != nil {
-		respondWithError(w, http.StatusInternalServerError, "Couldn't make JWT", err)
+		respond.WithError(w, http.StatusInternalServerError, "Couldn't make JWT", err)
 		return
 	}
 
 	refreshToken, err := auth.MakeRefreshToken()
 	if err != nil {
-		respondWithError(w, http.StatusInternalServerError, "Couldn't make refresh token", err)
+		respond.WithError(w, http.StatusInternalServerError, "Couldn't make refresh token", err)
 		return
 	}
 
-	_, err = cfg.db.CreateRefreshToken(r.Context(), database.CreateRefreshTokenParams{
+	_, err = a.DB.CreateRefreshToken(r.Context(), database.CreateRefreshTokenParams{
 		Token:     refreshToken,
 		UserID:    user.ID,
 		ExpiresAt: time.Now().Add(7 * 24 * time.Hour),
 	})
 	if err != nil {
-		respondWithError(w, http.StatusInternalServerError, "Couldn't save refresh token", err)
+		respond.WithError(w, http.StatusInternalServerError, "Couldn't save refresh token", err)
 		return
 	}
 
-	respondWithJSON(w, http.StatusCreated, response{
+	respond.WithJSON(w, http.StatusCreated, response{
 		User: User{
 			ID:        user.ID,
 			CreatedAt: user.CreatedAt,
